Always generate config files that don't exist
Automagically make copy of secret_config.py.template if secret_config.py
does not exist and generate secret key if it does not exist

+++ setup_app.py
@@ -1,38 +1,55 @@
 """ Script to set up application """
 
-# Parse arguments passed on command line
 import argparse
+import os
+
+# Parse arguments passed on command line
 parser = argparse.ArgumentParser()
-parser.add_argument("-q", "--quiet", 
-        help="Run quietly and force new secret key and db", 
-        action="store_true")
+parser.add_argument("-q", "--quiet",
+                    help="Run quietly and force new secret key and db",
+                    action="store_true")
 args = parser.parse_args()
 
-
-""" Generates a new secret key in config dir """
-def generate_new_secret():
-    import os
-    key_file_path = 'flask_app/config/secret_key.py'
-    secret_key = os.urandom(24).encode('hex').strip()
-    with open(key_file_path, 'w') as key_file:
-        key_file.write('""" Auto generated secret key file """\n\n' +
-                       'SECRET_KEY = "' + secret_key + '".decode("hex")\n')
+config_path = 'flask_app/config/'
 
 
-""" Initializes a new database """
+def generate_new_secret(force=False):
+    """ Generates a new secret key in config dir """
+    key_file_path = config_path + 'secret_key.py'
+    if force or not os.path.isfile(key_file_path):
+        secret_key = os.urandom(24).encode('hex').strip()
+        with open(key_file_path, 'w') as key_file:
+            key_file.write('""" Auto generated secret key file """\n\n' +
+                           'SECRET_KEY = "' + secret_key + '".decode("hex")\n')
+
+
+def copy_secret_config():
+    """ Make a copy of secret_config.py.template if secret_config.py does not exist """
+    if not os.path.isfile(config_path + 'secret_config.py'):
+        with open(config_path + 'secret_config.py', 'w') as secret_config_file:
+            with open(config_path + 'secret_config.py.template') as secret_config_file_template:
+                secret_config_file.write(secret_config_file_template.read())
+
+
 def initialize_db():
+    """ Initializes a new database """
     from flask_app.database import init_db
     init_db()
 
 
+copy_secret_config()
 if args.quiet:
     generate_new_secret()
     initialize_db()
 else:
-    input_answer = raw_input('Generate new secret key? (y, n)\n')
-    if input_answer.strip() == 'y':
+    if os.path.isfile(config_path + 'secret_key.py'):
+        # Ask for permission to overwrite
+        input_answer = raw_input('Generate new secret key and overwrite existing? (y, n)\n')
+        if input_answer.strip() == 'y':
+            generate_new_secret(True)
+    else:
         generate_new_secret()
 
-    input_answer = raw_input('Initialize database? (y, n)\n')
+    input_answer = raw_input('Initialize database (drops all tables)? (y, n)\n')
     if input_answer.strip() == 'y':
         initialize_db()
